+++ src/components/UI/Common/Title/Title.tsx
@@ -1,0 +1,15 @@
+import {ReactNode} from 'react'
+import styled from 'styled-components'
+
+const StyledTitleContainer = styled.h1`
+  margin: 20px 0;
+`
+
+const Title = ({children}: { children: ReactNode | string }) => {
+    return (
+        <StyledTitleContainer>
+            {children}
+        </StyledTitleContainer>
+    )
+}
+export default Title