+++ src/components/UI/Keypad/Key/Key.stories.tsx
@@ -1,9 +1,9 @@
-import { Story } from '@storybook/react'
+import {Story} from '@storybook/react'
 import Key, {KeyProps} from './Key'
 
 export default {
     component: Key,
-    title: 'Keypad/Key'
+    title: 'New/Keypad/Key'
 }
 
 const Template: Story<KeyProps> = (
@@ -17,7 +17,6 @@ const Template: Story<KeyProps> = (
 export const Primary = Template.bind({})
 
 Primary.args = {
-    number: 1,
-    alphabet: 'ABC',
-    onClick: () => console.log('onClick()')
+    number: 2,
+    alphabet: 'ABC'
 }