+++ src/components/UI/Keypad/Keypad.stories.tsx
@@ -1,0 +1,28 @@
+import {Story} from '@storybook/react'
+import Keypad, {KeypadProps} from './Keypad'
+
+export default {
+    component: Keypad,
+    title: 'New/Keypad/Keypad'
+}
+
+const Template: Story<KeypadProps> = (
+    args
+) => {
+    return (
+        <Keypad {...args}/>
+    )
+}
+
+export const Primary = Template.bind({})
+
+Primary.args = {
+    numberKeyOnClick: () => {
+    },
+    deleteKeyOnClick: () => {
+    },
+    cancelKeyOnClick: () => {
+    },
+    unlockKeyOnClick: () => {
+    }
+}