+++ src/components/UI/Layout/Page/Page.tsx
@@ -4,6 +4,8 @@ import {ReactNode} from "react";
 import StepButtons, {StepButtonsClickActions} from "../../StepButtons/StepButtons";
 import {pageTransitionVariants} from "../../Motion/Variants/Variants";
 import {screenWidth} from "../../screenSizes";
+import {LogoText} from "../../Logo/Logo";
+import {ChevronLeft} from "../../Icons/Icons";
 
 export type PageTransitionProps = {
     pageTransition: {
@@ -13,51 +15,55 @@ export type PageTransitionProps = {
 }
 
 type PageProps = {
-    titleComponent: ReactNode | string,
-    icon?: ReactNode,
     header?: ReactNode,
-    contentComponents: ReactNode,
+    contentComponents: {
+        contentLeft: ReactNode,
+        contentRight: ReactNode
+    },
     stepButtonsClickActions: StepButtonsClickActions
 } & PageTransitionProps
 
 
 const StyledPage = styled(motion.div)`
+  max-width: 1920px;
+  max-height: 1080px;
   width: 100%;
   height: 100%;
-  content: '';
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  background-color: ${props => props.theme.ui.backgroundColor};
-  position: absolute;
-  @media (${screenWidth.mediumWidth}) {
-    padding-top: 100px;
-  }
-`
-
-const StyledTitle = styled.h1`
-  width: 100%;
-  text-align: center;
-  color: ${props => props.theme.ui.textColor};
-  margin: 20px 0;
-`
-
-const StyledContents = styled.div`
-  max-width: 70rem;
-  width: 100%;
-  flex: 1;
-  height: 100%;
-  content: '';
-  color: ${props => props.theme.ui.textColor};
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  border-radius: 5px;
-  @media (${screenWidth.mediumWidth}) {
-    flex-direction: row;
+  position: absolute;
+  background-color: ${props => props.theme.ui.backgroundColor};
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.15);
+  @media (${screenWidth.narrowWidth}) {
+    padding: 20px 0;
   }
+`
+
+const StyledPageButton = styled.button`
+  width: 50px;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  background-color: transparent;
+  color: ${props => props.theme.ui.text.textPrimary}
+`
+
+const StyledContents = styled.div`
+  max-width: 1080px;
+  max-height: 1080px;
+  width: 100%;
+  height: 100%;
+  content: '';
+  color: ${props => props.theme.ui.text.textPrimary};
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
 `
 
 const StyledContent = styled.div`
@@ -65,24 +71,33 @@ const StyledContent = styled.div`
   height: 100%;
   display: flex;
   flex-direction: column;
-  align-items: center;
   justify-content: center;
+  align-items: center;
+  border-radius: 10px;
+  padding: 10px;
+  @media (${screenWidth.mediumWidth}) {
+    flex-direction: row;
+    height: 100%;
+  }
 `
 
-const StyledTitleContainer = styled(StyledContent)`
-  height: unset;
+
+const StyledContentTop = styled(StyledContent)`
+  flex-direction: column;
+  border: none;
+`
+
+const StyledContentBottom = styled(StyledContent)`
+  flex-direction: column;
+  border: none;
 `
 
 const StyledActionButtonContainer = styled.div`
-  max-width: 70em;
   width: 100%;
   margin-top: 30px
 `
 
 const Page = ({
-                  titleComponent,
-                  icon,
-                  header,
                   contentComponents,
                   stepButtonsClickActions,
                   pageTransition
@@ -100,19 +115,20 @@ const Page = ({
                 opacity: {duration: 0.3}
             }}>
             <StyledContents>
-                <StyledTitleContainer>
-                    {icon && icon}
-                    <StyledTitle>{titleComponent}</StyledTitle>
-                </StyledTitleContainer>
                 <StyledContent>
-                    {contentComponents}
+                    <StyledContentTop>
+                        {contentComponents.contentLeft}
+                    </StyledContentTop>
+                    <StyledContentBottom>
+                        {contentComponents.contentRight}
+                    </StyledContentBottom>
                 </StyledContent>
+                <StyledActionButtonContainer>
+                    {stepButtonsClickActions && (
+                        <StepButtons actionHandler={stepButtonsClickActions}/>
+                    )}
+                </StyledActionButtonContainer>
             </StyledContents>
-            <StyledActionButtonContainer>
-                {stepButtonsClickActions && (
-                    <StepButtons {...stepButtonsClickActions} />
-                )}
-            </StyledActionButtonContainer>
         </StyledPage>
     )
 }