+++ src/localization/messages/en-US.ts
@@ -2,8 +2,10 @@ import {Messages} from '../types'
 
 const en_US: Messages = {
     'main.greeting': 'Hello world!',
-    'onboarding.selectLanguageTitle': 'Select Language',
+    'onboarding.selectLanguageTitle': 'Select your language',
+    'onboarding.selectLanguageSubtitle': 'Select your language...',
     'onboarding.setPasscodeTitle': 'Create Passcode',
+    'onboarding.setPasscodeSubtitle': 'Enter a minimum 6 character passcode',
     'onboarding.confirmPasscodeTitle': 'Confirm Passcode',
     'onboarding.verificationTitle': 'Verify Account',
     'onboarding.verificationInputLabel': 'Enter your invitation code',